python-analyzer/requirements.txt
fastapi
uvicorn
httpx
pyyaml
apscheduler
pydantic
